+++ packages/terra-framework-docs/src/terra-dev-site/test/embedded-content-consumer/consumers/BasicConsumerWithNoScrolling.test.jsx
@@ -13,7 +13,9 @@ const BasicConsumerWithNoScrolling = () => (
     <h2>Embedded Content</h2>
     <p>The following is an embedded content within an iframe, no scrolling.</p>
     <EmbeddedContentConsumer
-      src="../../../../#/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/basic-provider"
+      // Use `../../../../` prefix if there is any changes to the providers file so we can see changes during the PR-preview build.
+      // Otherwise, keep the existing `/terra-framework/` prefix so the page will load when viewing the live site under /terra-ui and /terra-framework.
+      src="/terra-framework/#/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/basic-provider"
       className={cx('iframe')}
       options={{ iframeAttrs: { id: 'basic-consumer-frame' }, resizeConfig: { scrolling: false, fixedWidth: '100%', fixedHeight: '200px' } }}
       title="Basic Consumer"