+++ packages/terra-framework-docs/src/terra-dev-site/test/embedded-content-consumer/consumers/CustomEventsConsumer.test.jsx
@@ -49,7 +49,9 @@ class CustomEventsConsumer extends React.Component {
         <p>The following is an embedded content within an iframe.</p>
         <div id="CustomEvents">
           <EmbeddedContentConsumer
-            src="../../../../#/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/custom-events-provider"
+            // Use `../../../../` prefix if there is any changes to the providers file so we can see changes during the PR-preview build.
+            // Otherwise, keep the existing `/terra-framework/` prefix so the page will load when viewing the live site under /terra-ui and /terra-framework.
+            src="/terra-framework/#/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/custom-events-provider"
             className={cx('iframe')}
             title="Custom events example"
             options={{ iframeAttrs: { id: 'custom-events-consumer-frame' }, resizeConfig: { scrolling: true } }}